+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [unreleased]
+
+### Changed
+
+- added `match` to the return object from `convertLetterSidc2NumberSidc`. Possible values are currently `exact`
+  and `failed`.
+
 ## [0.2.5] - 2023-03-19
 
 ### Fixed